requirements.txt
psycopg2-binary==2.9.5
pyyaml==6.0
schedule==1.1.0
python-telegram-bot==13.14
requests==2.28.2
apscheduler==3.10.1
psutil==5.9.4
flask==2.2.3
prometheus-client==0.16.0
SQLAlchemy==2.0.4
python-dotenv==1.0.0 